Load rails before parsing queues Configuration loading files like dotenv and figaro need to be loaded before Shoryuken parses the shoryuken config file, which may contain environment variables
