+++ lib/shoryuken/runner.rb
@@ -28,6 +28,7 @@ module Shoryuken
         end
       end
 
+      load_rails if options[:rails]
       loader = EnvironmentLoader.setup_options(options)
 
       # When cli args exist, override options in config file
@@ -133,6 +134,26 @@ module Shoryuken
         logger.info { "Received #{sig}, will shutdown" }
 
         raise Interrupt
+      end
+    end
+
+    def load_rails
+      # Adapted from: https://github.com/mperham/sidekiq/blob/master/lib/sidekiq/cli.rb
+
+      require 'rails'
+      if ::Rails::VERSION::MAJOR < 4
+        require File.expand_path('config/environment.rb')
+        ::Rails.application.eager_load!
+      else
+        # Painful contortions, see 1791 for discussion
+        require File.expand_path('config/application.rb')
+        if ::Rails::VERSION::MAJOR == 4
+          ::Rails::Application.initializer 'shoryuken.eager_load' do
+            ::Rails.application.config.eager_load = true
+          end
+        end
+        require 'shoryuken/extensions/active_job_adapter' if Shoryuken.active_job?
+        require File.expand_path('config/environment.rb')
       end
     end
   end
